+++ inst/include/rxode2_model_shared.c
@@ -207,6 +207,36 @@ double _sign(unsigned int n, ...) {
   return s;
 }
 
+double _mix(int _cSub, unsigned int n,  ...) {
+  rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
+  va_list valist;
+  va_start(valist, n);
+  double ret = NA_REAL;
+  double p = 0.0;
+  double u = ind->mixunif; // sampled once per patient
+  int found = 0;
+  double v = 0.0;
+  for (unsigned int i = 0; i < n-1; i+= 2) {
+    v = va_arg(valist, double);
+    p += va_arg(valist, double);
+    if (found == 0) {
+      if (u < p) {
+        ret = v;
+        ind->mixest = i/2 + 1; // 1-based index
+        found = 1;
+      }
+    }
+  }
+  v = va_arg(valist, double); // other possibility
+  if (found == 0) {
+    ind->mixest = _solveData->mixnum;
+    ret = v;
+  }
+  va_end(valist);
+  return ret;
+}
+
+
 double _rxord(int _cSub, unsigned int n,  ...) {
   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
   if (!ind->inLhs) {