+++ src/rx2api.c
@@ -67,6 +67,29 @@ int getIndIx(rx_solving_options_ind* ind, int j) {
   return ind->ix[j];
 }
 
+int getIndMixest(rx_solving_options_ind* ind) {
+  return ind->mixest;
+}
+
+void setIndMixest(rx_solving_options_ind* ind, int mixest) {
+  if (mixest < 0) {
+    Rf_error("[setIndMixest]: mixest (%d) should be >= 0", mixest);
+  }
+  ind->mixest = mixest;
+}
+
+int getRxMixnum(rx_solve *rx) {
+  // This is the number of mixtures
+  return rx->mixnum;
+}
+
+void setRxMixnum(rx_solve *rx, int mixnum) {
+  if (mixnum < 0) {
+    Rf_error("[setRxMixnum]: mixnum (%d) should be >= 0", mixnum);
+  }
+  rx->mixnum = mixnum;
+}
+
 int getIndEvid(rx_solving_options_ind* ind, int kk) {
   if (kk < 0 || kk >= ind->n_all_times) {
     Rf_error("[getIndEvid]: kk (%d) should be between [0, %d)", kk, ind->n_all_times);