+++ src/tran.c
@@ -425,6 +425,8 @@ void reset(void) {
   tb.nInd       = 0;
   tb.simflg     = 0;
   tb.nLlik      = 0;
+  tb.hasMix     = 0;
+
   // Reset Arrays
   // Reset integers
   NV		= 0;